push_ad/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
db-dtypes==1.0.5
Django==4.1.4
django-filter==22.1
djangorestframework==3.14.0
google-ads==19.0.0
google-api-core==2.10.1
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==2.30.1
google-cloud-core==2.3.2
google-cloud-monitoring==1.1.2
google-cloud-trace==0.24.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.48.2
idna==3.4
importlib-metadata==5.2.0
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.24.0
oauthlib==3.2.2
opentelemetry-api==0.17b0
opentelemetry-exporter-google-cloud==0.18b1
opentelemetry-instrumentation==0.17b0
opentelemetry-sdk==0.17b0
packaging==21.3
pandas==1.5.2
proto-plus==1.22.1
# protobuf==4.21.12
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
wrapt==1.14.1
zipp==3.11.0
